Switched source.on_track(fun) to on_track(fun, source)
This is a change in syntax between the current version of liquidsoap and
the developer branch. This code was originally written for the
developer branch, but is now being modified to run on the more stable
v1.4.4.

I suspect it will need to be changed back when the liquidsoap developers
move this syntax change back into the main branch, although I have no
idea when that will be.

`on_track` (v1.4.4) reference:
https://www.liquidsoap.info/doc-1.4.4/reference.html#on_track

`source.on_track` (dev) reference:
https://www.liquidsoap.info/doc-dev/reference.html#source.on_track